A few trait fixes (#16062)
